Move AttributeTemplate to rustc_attr_parsing
